requirements.txt
opencv-python>=4.11.0.86
numpy>=1.26.4
pillow>=10.4.0
pywin32>=306; platform_system == "Windows"
